requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
boto==2.46.1
boto3==1.4.4
botocore==1.5.78
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-boto==0.3.9
django-cors-headers==1.1.0
django-debug-toolbar==1.7
django-downloadview==1.9
django-filter==1.0.2
django-mail-queue==2.2.2
django-ordered-model==1.4.1
django-rest-auth==0.9.1
django-storages-redux==1.3.2
djangorestframework==3.6.2
docutils==0.13.1
drf-nested-routers==0.90.0
futures==3.1.1
httplib2==0.10.3
idna==2.5
image==1.5.5
isoweek==1.3.0
jmespath==0.9.3
Markdown==2.6.8
markdown2==2.3.3
numpy==1.13.1
oauth2==1.9.0.post1
oauthlib==0.1.3
olefile==0.44
packaging==16.8
parse==1.8.2
pathlib==1.0.1
Pillow==4.2.0
psycopg2==2.7.1
pyapi-gitlab==7.8.5
pyasn1==0.2.3
pyparsing==2.2.0
python-dateutil==2.6.0
python-gitlab==0.21.2
pytz==2017.2
PyVimeo==0.3.10
requests==2.18.1
rsa==3.4.2
s3transfer==0.1.10
simplejson==3.11.1
six==1.10.0
sqlparse==0.2.3
upsmychoice==1.0.2
urllib3==1.21.1
vimeo==0.1.9
